fix delete post bug

+++ src/public/js/post-manage.js
@@ -125,35 +125,48 @@ $(function(){
 
 //DELETE POST
 $(function(){
-    $("form.delete-form").click("submit", function(e){
+    $("form.delete-form").on("submit", function(e){
         e.preventDefault();
-        var dataStr = $(this).serialize();
-        // console.log(dataStr_arr);
+
+        Swal.fire({
+            title: 'Bạn có chắc muốn xóa bài này?',
+            text: "Sau khi xóa, bài viết không thể khôi phục lại",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Xóa luôn!'
+            }).then((result) => {
+                if (result.isConfirmed) {
+                    var dataStr = $(this).serialize();
+                    var callback = $.ajax({
+                        type: "DELETE",
+                        url: "/post",
+                        data: dataStr,
+                        success: ()=>{
+                            // console.log("Update pop:", dataStr);
+                            var headerdata = callback.getResponseHeader('signal');
+                            if (headerdata == 1){
+                                Swal.fire(
+                                    'Đã xóa',
+                                    'Qua trang chủ kiểm tra được chưa mate',
+                                    'success'
+                                ).then(()=>{
+                                    window.location.reload();
+                                })
+                            }
+                            else{
+                                Swal.fire(
+                                    'Chưa xóa',
+                                    'Lỗi gì gòi, hỏi thằng làm web xem',
+                                    'error'
+                                )
+                            }
+                        }
+                    })   
+                }
+            })
+            
         
-        // console.log(dataStr);
-            // console.log(dataStr);
-        var callback = $.ajax({
-            type: "DELETE",
-            url: "/post",
-            data: dataStr,
-            success: ()=>{
-                // console.log("Update pop:", dataStr);
-                var headerdata = callback.getResponseHeader('signal');
-                if (headerdata == 1){
-                    Swal.fire(
-                        'Đã xóa',
-                        'Qua trang chủ kiểm tra được chưa mate',
-                        'success'
-                    )
-                }
-                else{
-                    Swal.fire(
-                        'Chưa xóa',
-                        'Lỗi gì gòi, hỏi thằng làm web xem',
-                        'error'
-                    )
-                }
-            }
-        })   
     })
 });
